Tinh tong 2 so (doi float sang int)

+++ main.py
@@ -2,8 +2,8 @@ def sum_two_numbers(a, b):
     return a + b
 
 def main():
-    a = float(input("Nhập số a: "))
-    b = float(input("Nhập số b: "))
+    a = int(input("Nhập số a: "))
+    b = int(input("Nhập số b: "))
     
     result = sum_two_numbers(a, b)
     print("Tổng của a và b là:", result)
